Extract JSON parsers and NumpyEncoder into parsers module
Move the six small JSON / dimension-universe helpers out of `utils.py`
into a new `parsers` module:

- NumpyEncoder — JSON encoder that knows how to render numpy ints,
  floats and ndarrays
- sanitizeNans — recursive replacement of NaN with None and coercion
  of numeric strings to floats
- safeJsonOpen — JSON loader that retries through a short timeout to
  tolerate concurrent writers
- writeDimensionUniverseFile / getDimensionUniverse — read/write the
  per-location dimension universe JSON file used to round-trip
  DimensionRecords through Redis
- expRecordFromJson — DimensionRecord deserialiser that uses the
  dimension universe loaded above

These six helpers form the third carve-out of the utils.py split.
They have no internal cross-dependencies inside `utils.py` and only
the dimension-universe helpers reference `LocationConfig` (via a
TYPE_CHECKING import that still points at `.utils` for now — it will
be re-pointed once `LocationConfig` itself moves).

`utils.py` continues to use `NumpyEncoder` internally inside
`writeMetadataShard` and `writeDataShard`, so it now reaches back
into the new module via `from .parsers import NumpyEncoder`. The
previously-unused `math`, `DimensionConfig` and `DimensionUniverse`
imports in `utils.py` have been removed.

All callers across the package, scripts, and tests have been updated
to import from `parsers` instead of `utils`. As per the
"this is an application, not a library" convention, no
compatibility shims are added.

Verified:
- mypy and flake8 clean across the package, scripts, and tests.
- The full unit test suite (185 tests) passes.
- Every dependent in-package module imports cleanly.

Written by Claude Code 4.6 Opus (high-effort).
