+++ python/lsst/rubintv/production/utils.py
@@ -25,7 +25,6 @@ import glob
 import io
 import json
 import logging
-import math
 import os
 import sys
 import time
@@ -43,15 +42,14 @@ import yaml
 from lsst.daf.butler import (
     Butler,
     DataCoordinate,
-    DimensionConfig,
     DimensionGroup,
     DimensionRecord,
-    DimensionUniverse,
 )
 from lsst.obs.lsst.translators.lsst import FILTER_DELIMITER
 from lsst.resources import ResourcePath
 from lsst.utils import getPackageDir
 
+from .parsers import NumpyEncoder
 from .predicates import isFileWorldWritable, runningScons
 
 if TYPE_CHECKING:
@@ -62,21 +60,16 @@ if TYPE_CHECKING:
 
 __all__ = [
     "setupSentry",
-    "writeDimensionUniverseFile",
-    "getDimensionUniverse",
     "checkRubinTvExternalPackages",
     "writeMetadataShard",
     "writeDataShard",
     "getShardedData",
     "LocationConfig",
     "getAutomaticLocationConfig",
-    "sanitizeNans",
-    "safeJsonOpen",
     "getNumExpectedItems",
     "getShardPath",
     "removeDetector",
     "ALLOWED_DATASET_TYPES",
-    "NumpyEncoder",
     "logDuration",
     "timeFunction",
     "summaryStatsToDict",
@@ -118,21 +111,6 @@ def setupSentry() -> None:
         logger.warning("Sentry DSN not found or client inactive — events will not be reported")
 
 
-def writeDimensionUniverseFile(butler, locationConfig: LocationConfig) -> None:
-    """Run on butler watcher startup.
-
-    This assumes that all repos in a give location are on the same version, but
-    we will make sure to keep that always true.
-    """
-    with open(locationConfig.dimensionUniverseFile, "w") as f:
-        f.write(json.dumps(butler.dimensions.dimensionConfig.toDict()))
-
-
-def getDimensionUniverse(locationConfig: LocationConfig) -> DimensionUniverse:
-    duJson = safeJsonOpen(locationConfig.dimensionUniverseFile)
-    return DimensionUniverse(DimensionConfig(duJson))
-
-
 def getGlobPatternForShardedData(
     path: str, dataSetName: str, instrument: str, dayObs: int, seqNum: int
 ) -> str:
@@ -163,19 +141,6 @@ def getGlobPatternForShardedData(
         seqNum=seqNumFormatted,
         suffix="*",
     )
-
-
-def expRecordFromJson(expRecordJson: str | bytes, locationConfig: LocationConfig) -> DimensionRecord | None:
-    """Deserialize a DimensionRecord from a JSON string.
-
-    expRecordJson : `str` or `bytes`
-        The JSON string to deserialize, as either a string or bytes.
-    locationConfig : `lsst.rubintv.production.utils.LocationConfig`
-        The location configuration, used to determine the dimension universe.
-    """
-    if not expRecordJson:
-        return None
-    return DimensionRecord.from_json(expRecordJson, universe=getDimensionUniverse(locationConfig))
 
 
 @dataclass(frozen=True)
@@ -666,61 +631,6 @@ def catchPrintOutput(functionToCall: Callable, *args, **kwargs) -> str:
     return f.getvalue()
 
 
-def sanitizeNans(obj: Any) -> Any:
-    """
-    Recursively sanitize an object by replacing NaN values with None and coerce
-    numeric-looking strings to floats.
-
-    Nans are not JSON de-serializable, so this function replaces them with
-    ``None``. In addition, strings that represent numeric values (including
-    scientific notation) are converted to floats to ensure proper JSON numeric
-    typing.
-
-    Parameters
-    ----------
-    obj : `object`
-        The object to santitize, expected to be dict-like: either a single
-        ``dict`` or a ``list`` or ``dict`` of ``dict``s.
-
-    Returns
-    -------
-    obj : `object`
-        The object with any NaNs replaced with ``None``, and numeric strings
-        converted to floats.
-    """
-    if isinstance(obj, list):
-        return [sanitizeNans(o) for o in obj]
-    elif isinstance(obj, dict):
-        return {k: sanitizeNans(v) for k, v in obj.items()}
-    elif isinstance(obj, float) and math.isnan(obj):
-        return None
-    elif isinstance(obj, str):
-        s = obj.strip()
-        if not s:
-            return obj
-        # Try to coerce numeric-looking strings to floats; map NaN to None.
-        try:
-            value = float(s)
-            if math.isnan(value):
-                return None
-            return value
-        except Exception:
-            return obj
-    else:
-        return obj
-
-
-class NumpyEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, np.integer):
-            return int(obj)
-        elif isinstance(obj, np.floating):
-            return float(obj)
-        elif isinstance(obj, np.ndarray):
-            return obj.tolist()
-        return super().default(obj)
-
-
 def writeMetadataShard(path: str, dayObs: int, mdDict: dict[int, dict[str, Any]]) -> None:
     """Write a piece of metadata for uploading to the main table.
 
@@ -991,43 +901,6 @@ def getShardedData(
         if deleteRegardless or (deleteIfComplete and len(files) == nExpected):
             os.remove(dataShard)
     return data, len(files)
-
-
-def safeJsonOpen(filename: str, timeout=0.3) -> str:
-    """Open a JSON file, waiting for it to be populated if necessary.
-
-    JSON doesn't like opening zero-byte files, so try to open it, and if it's
-    empty, add a series of small waits until it's not empty and reads
-    correctly, or the timeout is reachecd.
-
-    Parameters
-    ----------
-    filename : `str`
-        The filename to open.
-    timeout : `float`, optional
-        The timeout period after which to give up waiting for the contents to
-        populate.
-
-    Returns
-    -------
-    jsonData : `str`
-        The data from the json file as a string, i.e. not put back into a
-        python object.
-
-    Raises
-    ------
-    RuntimeError: Raised if the file is not populated within the timeout.
-    """
-    start = time.time()
-    while time.time() - start < timeout:
-        try:
-            with open(filename, "r") as f:
-                jsonData = json.load(f)
-                return jsonData
-        except (RuntimeError, json.decoder.JSONDecodeError):
-            pass
-        time.sleep(0.1)
-    raise RuntimeError(f"Failed to load data from {filename} after {timeout}s")
 
 
 def getNumExpectedItems(expRecord: DimensionRecord, logger: Logger | None = None) -> int:
