+++ scripts/LSSTCam/runButlerWatcher.py
@@ -21,12 +21,9 @@
 
 from lsst.daf.butler import Butler
 from lsst.rubintv.production import ButlerWatcher
+from lsst.rubintv.production.parsers import writeDimensionUniverseFile
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import (
-    getAutomaticLocationConfig,
-    setupSentry,
-    writeDimensionUniverseFile,
-)
+from lsst.rubintv.production.utils import getAutomaticLocationConfig, setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
