+++ scripts/LSSTComCamSim/runButlerWatcher.py
@@ -21,8 +21,9 @@
 
 from lsst.daf.butler import Butler
 from lsst.rubintv.production import ButlerWatcher
+from lsst.rubintv.production.parsers import writeDimensionUniverseFile
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import getAutomaticLocationConfig, writeDimensionUniverseFile
+from lsst.rubintv.production.utils import getAutomaticLocationConfig
 from lsst.summit.utils.utils import setupLogging
 
 setupLogging()
